airflow_retail_pg/requirements.txt
astronomer-cosmos
apache-airflow-providers-postgres